# Add options to ignore default values during serialization