--- a/src/libraries/System.Text.Json/ref/System.Text.Json.cs
+++ b/src/libraries/System.Text.Json/ref/System.Text.Json.cs
@@ -213,6 +213,7 @@ namespace System.Text.Json
         public bool AllowTrailingCommas { get { throw null; } set { } }
         public System.Collections.Generic.IList<System.Text.Json.Serialization.JsonConverter> Converters { get { throw null; } }
         public int DefaultBufferSize { get { throw null; } set { } }
+        public System.Text.Json.Serialization.JsonIgnoreCondition DefaultIgnoreCondition { get { throw null; } set { } }
         public System.Text.Json.JsonNamingPolicy? DictionaryKeyPolicy { get { throw null; } set { } }
         public System.Text.Encodings.Web.JavaScriptEncoder? Encoder { get { throw null; } set { } }
         public bool IgnoreNullValues { get { throw null; } set { } }
@@ -465,9 +466,9 @@ namespace System.Text.Json.Serialization
 {
     public enum JsonIgnoreCondition
     {
-        Always = 0,
-        WhenNull = 1,
-        Never = 2,
+        Never = 0,
+        Always = 1,
+        WhenWritingDefault = 2,
     }
     public abstract partial class JsonAttribute : System.Attribute
     {