--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonResumableConverterOfT.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonResumableConverterOfT.cs
@@ -41,6 +41,6 @@ namespace System.Text.Json.Serialization
             TryWrite(writer, value, options, ref state);
         }
 
-        public override bool HandleNull => false;
+        public sealed override bool HandleNull => false;
     }
 }