--- a/src/libraries/System.Text.Json/tests/JsonTestHelper.cs
+++ b/src/libraries/System.Text.Json/tests/JsonTestHelper.cs
@@ -868,10 +868,14 @@ namespace System.Text.Json
                     Assert.Equal(expected.GetString(), actual.GetString());
                     break;
                 case JsonValueKind.Number:
+                case JsonValueKind.True:
+                case JsonValueKind.False:
+                case JsonValueKind.Null:
                     Assert.Equal(expected.GetRawText(), actual.GetRawText());
                     break;
                 default:
-                    throw new NotImplementedException();
+                    Debug.Fail($"Unexpected JsonValueKind: JsonValueKind.{valueKind}.");
+                    break;
             }
         }
     }