… src/libraries/System.Text.Json/tests/Serialization/OptionsTests.cs
… src/libraries/System.Text.Json/tests/Serialization/PropertyVisibilityTests.cs