Fixes https://github.com/dotnet/runtime/issues/779.

Breaking changes

- [between .NET 5 preview 5 & 6] `JsonIgnoreCondition.WhenNull` removed & replaced with `JsonIgnoreCondition.WhenWritingDefault`
- [between .NET Core 3.x and .NET 5] For edge case scenarios where a converter returns `null` when the token type is not `JsonTokenType.Null`, the property's setter is called (with the null arg). The setter was not called before.