@layomia 2020-05-13
<details>
<summary>
This change does not regress existing benchmarks (click for summary).
</summary>

<br/>

```
dotnet run --base "D:\benchmarks\default\before" --diff "D:\benchmarks\default\after" --threshold 2%
summary:
better: 2, geomean: 1.029
worse: 2, geomean: 1.030
total diff: 4

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.WriteJson<IndexViewModel>.SerializeToUtf8By |      1.03 |         17152.36 |         17669.03 |         |
| System.Text.Json.Serialization.Tests.WriteJson<BinaryData>.SerializeToUtf8Bytes  |      1.03 |           442.96 |           456.19 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.ReadJson<BinaryData>.DeserializeFromUtf8Byt |      1.03 |           535.32 |           520.09 |         |
| System.Text.Json.Serialization.Tests.WriteJson<SimpleStructWithProperties>.Seria |      1.03 |           269.63 |           262.07 |         |
```
</details>

---

A simple benchmark comparing the serialization of a POCO with 6 value-type properties with and without "ignore default value" semantics showed 20% faster serialization & 45% decrease in allocs with when ignoring defaults.

|                 Method |     Mean |    Error |   StdDev |   Median |      Min |      Max | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------:|-------:|------:|------:|----------:|
| SerializeDefaultValues | 648.9 ns | 12.42 ns | 11.61 ns | 643.6 ns | 636.9 ns | 669.1 ns |  1.00 | 0.1463 |     - |     - |     616 B |
| IgnoreDefaultValues    | 536.1 ns | 7.41 ns | 6.57 ns | 537.4 ns | 523.6 ns | 545.4 ns   |  1.00 | 0.0802 |     - |     - |     336 B |

I plan to benchmark more scenarios when this PR has been picked up by dotnet/performance, so I can easily use the new API.

<details>
<summary>
More on benchmark (click to view)
</summary

<br/>
<br/>

The following benchmark was run using two versions of dotnet/runtime:
- a build of baseline master (SerializeDefaultValues )
- a build containing the changes in this PR (IgnoreDefaultValues ), but with the option to ignore default values hardcoded true (to bypass new API that hasn't been picked up by the dotnet/performance repo). This detail doesn't affect the benchmark results.

The results for each run were combined into the table above.

```cs
// Licensed to the .NET Foundation under one or more agreements.
// The .NET Foundation licenses this file to you under the MIT license.
// See the LICENSE file in the project root for more information.

using BenchmarkDotNet.Attributes;
using MicroBenchmarks;
using MicroBenchmarks.Serializers;

namespace System.Text.Json.Serialization.Tests
{
    [GenericTypeArguments(typeof(SimpleClassWithValueTypeProperties))]
    public class IgnoreDefault<T>
    {
        private T _value;

        [GlobalSetup]
        public void Setup()
        {
            _value = DataGenerator.Generate<T>();
        }

        [BenchmarkCategory(Categories.Libraries, Categories.JSON)]
        [Benchmark(Baseline = true)]
        public byte[] SerializeDefaultValues() => JsonSerializer.SerializeToUtf8Bytes(_value);
    }
}
```

Test POCOS

```cs
public class SimpleClassWithValueTypeProperties
{
    public int Int { get; set; }
    public float Float { get; set; }
    public decimal Decimal { get; set; }
    public bool Bool { get; set; }
    public Guid Guid { get; set; }
    public SimpleStructWithProperties SimpleStruct { get; set; }
}

public struct SimpleStructWithProperties
{
    public int Num { get; set; }
    public string Text { get; set; }
}
```

</details>

[after.zip](https://github.com/dotnet/runtime/files/4619819/after.zip)
[before.zip](https://github.com/dotnet/runtime/files/4619820/before.zip)
[ignore_default.zip](https://github.com/dotnet/runtime/files/4619821/ignore_default.zip)
[serialize_default.zip](https://github.com/dotnet/runtime/files/4619822/serialize_default.zip)