@danmosemsft 2020-07-24
@layomia this is labeled breaking change. Can you please open an issue with https://github.com/dotnet/docs/issues/new?template=dotnet-breaking-change.md 

I don't [see an existing one](https://github.com/dotnet/docs/issues?q=+is%3Aissue+label%3Abreaking-change+)