@layomia 2020-10-03
> - [between .NET 5 preview 5 & 6] `JsonIgnoreCondition.WhenNull` removed & replaced with `JsonIgnoreCondition.WhenWritingDefault`

This was a breaking change between previews, so a breaking change doc is not needed.

> - [between .NET Core 3.x and .NET 5] For edge case scenarios where a converter returns `null` when the token type is not `JsonTokenType.Null`, the property's setter is called (with the null arg). The setter was not called before.

This is actually not true, so this PR did **not** introduce a breaking change. See this example (also on [dotnetfiddle](https://dotnetfiddle.net/nmT96K))

```cs
using System;
using System.Text.Json;
using System.Text.Json.Serialization;
					
public class Program
{
	public static void Main()
	{
		var options = new JsonSerializerOptions
		{ 
			Converters = { new StringConverter() },
			IgnoreNullValues = true
		};
		
		string json = @"{""MyString"":null}";
		ClassWithString obj = JsonSerializer.Deserialize<ClassWithString>(json, options);
		Console.WriteLine(obj.MyString);
		// .NET 3.1: "Random"
		// .NET 5.0: "Random"
		
		json = @"{""MyString"":""Hello World""}";
		obj = JsonSerializer.Deserialize<ClassWithString>(json, options);
		Console.WriteLine(obj.MyString);
		// .NET 3.1: null
		// .NET 5.0: null
	}
	
	private class ClassWithString
	{
		public string MyString { get; set; } = "Random";
	}
	
	private class StringConverter : JsonConverter<string>
	{
		public override string Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
		{
			return null;
		}
		
		public override void Write(Utf8JsonWriter writer, string value, JsonSerializerOptions options) =>
			throw new NotImplementedException();
	}
}
```